refactor: move CsrfGuard to src/common/guards per architecture rules
