+++ src/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { AuthRepository } from './repositories/auth.repository';
 import { SupabaseAuthGuard } from './guards/supabase-auth.guard';
-import { CsrfGuard } from './guards/csrf.guard';
+import { CsrfGuard } from '../common/guards/csrf.guard';
 import { UsersModule } from '../modules/users/users.module';
 import { ObservabilityModule } from '../observability/observability.module';
